+++ js/data.js
@@ -1,0 +1,20 @@
+const FOR_TYPE_MAPS = {
+  1: 'palace',
+  2: 'flat',
+  3: 'house',
+  4: 'bungalow',
+};
+const FOR_CHECK_IN = {
+  1: '12:00',
+  2: '13:00',
+  3: '14:00',
+};
+const FOR_FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+export {FOR_TYPE_MAPS, FOR_CHECK_IN, FOR_FEATURES};
